requirements.txt
# 核心依赖
playwright==1.49.0
pytest==8.3.4
pytest-playwright==0.5.2
allure-pytest==2.13.5
python-dotenv==1.0.1

# 测试增强
pytest-xdist==3.5.0           # 并行测试执行
pytest-rerunfailures==14.0    # 失败重试机制
pytest-timeout==2.3.1         # 测试超时控制

# 代码质量
ruff==0.4.0                   # 代码质量检查和格式化
